+++ src/parser/gpx.rs
@@ -1,10 +1,15 @@
 //! gpx handles parsing of GPX elements.
 
-use std::io::Read;
-
+#[cfg(feature = "time")]
+use ::time::OffsetDateTime as Time;
+#[cfg(feature = "chrono")]
 use chrono::{DateTime, Utc};
 use geo_types::Rect;
+use std::io::Read;
 use xml::reader::XmlEvent;
+
+#[cfg(not(feature = "time"))]
+type Time = DateTime<Utc>;
 
 use crate::errors::{GpxError, GpxResult};
 use crate::parser::{
@@ -29,7 +34,7 @@ pub fn consume<R: Read>(context: &mut Context<R>) -> Result<Gpx, GpxError> {
     let mut url: Option<String> = None;
     let mut urlname: Option<String> = None;
     let mut email: Option<String> = None;
-    let mut time: Option<DateTime<Utc>> = None;
+    let mut time: Option<Time> = None;
     let mut bounds: Option<Rect<f64>> = None;
     let mut gpx_name: Option<String> = None;
     let mut description: Option<String> = None;