+++ src/writer.rs
@@ -2,9 +2,15 @@
 
 use std::io::Write;
 
+#[cfg(feature = "chrono")]
 use chrono::{DateTime, Utc};
 use geo_types::Rect;
+#[cfg(feature = "time")]
+use time::{format_description::well_known::Rfc3339, OffsetDateTime as Time};
 use xml::writer::{EmitterConfig, EventWriter, XmlEvent};
+
+#[cfg(not(feature = "time"))]
+type Time = DateTime<Utc>;
 
 use crate::errors::{GpxError, GpxResult};
 use crate::types::*;
@@ -238,12 +244,16 @@ fn write_person_if_exists<W: Write>(
 }
 
 fn write_time_if_exists<W: Write>(
-    time: &Option<DateTime<Utc>>,
+    time: &Option<Time>,
     writer: &mut EventWriter<W>,
 ) -> GpxResult<()> {
     if let Some(ref time) = time {
         write_xml_event(XmlEvent::start_element("time"), writer)?;
+        #[cfg(feature = "time")]
+        write_xml_event(XmlEvent::characters(&time.format(&Rfc3339)?), writer)?;
+        #[cfg(feature = "chrono")]
         write_xml_event(XmlEvent::characters(&time.to_rfc3339()), writer)?;
+
         write_xml_event(XmlEvent::end_element(), writer)?;
     }
     Ok(())